:art: rename topup directory
